requirements.txt
asgiref==3.8.1
Django==5.0.6
djangorestframework==3.15.1
djangorestframework-simplejwt==5.3.1
pip==24.0
psycopg2-binary==2.9.9
PyJWT==2.8.0
python-decouple==3.8
setuptools==65.5.0
sqlparse==0.5.0